requirements.txt
openpyxl
mysql-connector-python ## CONNECT BDD
beautifulsoup4 ## HTML PARSING
requests ## GET DATA FROM URL